+++ build.gradle
@@ -12,9 +12,7 @@ repositories {
 dependencies {
     compile gradleApi()
     compile localGroovy()
-    compile 'org.codehaus.groovy.modules.http-builder:http-builder:0.7.1'
     compile 'org.apache.httpcomponents:httpmime:4.2.5'
-    runtime 'org.codehaus.groovy.modules.http-builder:http-builder:0.7.1'
     runtime 'org.apache.httpcomponents:httpmime:4.2.5'
 }
