+++ src/main/groovy/com/deploygate/gradle/plugins/Config.groovy
@@ -1,6 +1,5 @@
 package com.deploygate.gradle.plugins
 
 class Config {
-    static final def DEPLOYGATE_HOST = 'https://deploygate.com'
-    static final def API_END_POINT = "${DEPLOYGATE_HOST}/api"
+    static final def DEPLOYGATE_ROOT = 'https://deploygate.com'
 }
